The Wall

+++ _python/django/Full Stack Django/the_wall_project/the_wall_app/views.py
@@ -1,15 +1,23 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.utils import timezone
+from datetime import datetime
 from the_wall_app.models import *
 from login_and_regestration_app.models import *
 
 def wall(request):
     if "status" in request.session:
         if request.session["status"] == "Logged In":
+            for message in Message.objects.all():
+                if (timezone.now()- message.created_at).total_seconds()/60 > 30:
+                    print(message.deletable)
+                    message.deletable = False
+                    message.save()
             context = {
                 "user": User.objects.get(id=request.session["userid"]),
                 "messages": Message.objects.all(),
                 "comments": Comment.objects.all(),
+                "timenow": timezone.now(),
             }
             return render(request, 'wall.html', context)
     return redirect("/")
@@ -26,7 +34,7 @@ def create_comment(request, id):
 
 def delete(request, id):
     this_message = Message.objects.get(id = id)
-    if request.session["userid"] == this_message.user.id:
+    if request.session["userid"] == this_message.user.id and this_message.deletable:
         this_message.delete()
 
     return redirect('/wall')
